Get basic form functionality working
- Add IDs to username/password fields (for easy styling, honestly)
- Make button a "submit" button
- Remove `console.log()` since the values were now being submitted
- Reset the form after the username and password are collected
- Set the username and password vars to 'undefined'

Changing the button type to 'submit' was necessary to make the form
actually submit, since the event handler is for a form submitting
instead of a button click.

Also, setting the username and password vars to `undefined` is just to
remove any trace in memory of the values that were set -- just in case.