Get projects manager to display current project info (issue #3)
All project info that needs to be edited is displayed in form fields.
This will be useful if the fields need to be updated later, and these
fields can be styled easily.

Next features that need to be added to this:
- Allow updated existing projects
- Allow adding new projects
- Input security

Maybe for adding projects, the project document can be initialized with
empty/default fields as soon as the "add" button is clicked. Then the
"updating" part will work like normal.