Fix validation issue and return result object
The fix is in (hopefully): turns out the document ID needs to be a
string 17 characters long, not an `ObjectId` object containging a hex
string 24 characters long. That's what `SimpleSchema.RegEx.Id` is used
to check for.

I also setup a better `if...else` to return an object with either an
error or success message so it can be used client-side.