Fix validation issues and give methods some breathing room
The validation issue ended up being caused by a dumb mistake.

This is what the issue was: I was passing a variable to
`idContext.validate()` that had a different name than the key I was
trying to validate for: the `_id` key. So it was, correctly, looking for
a key named `projID` to validate my `projID` variable against.

To fix this issue, I changed it to explicitly set the key `_id` to the
value of `projID` before validating.

I'm also currently using a simpler regex since I'm not sure if the
SimpleSchema package author intended to leave out the characters
`01IOUVl`, or if he just mistyped while he was making the regex:
```
/^[23456789ABCDEFGHJKLMNPQRSTWXYZabcdefghijkmnopqrstuvwxyz]{17}$/
```

The above regex leaves out the characters `01IOUVl`. So I'm currently
using the regex `/^[0-9A-Za-z]{17}$/` since it's much simpler.

If I find out that the regex stored at `SimpleSchema.RegEx.Id` was
intentional, then I'll switch back.