Fix issue causing undefined responses
I wrote the server-side method to expect `update` to be an object with a property `docsAffected`. This was causing `success` on the client to be undefined.

This change will fix this issue.

Also, add a note to be sure `update` will only ever contain a number.

According to [these docs][1] on updates, the alternative is that there will be an error thrown. But I'll look into it more later to be sure this is entirely the expected behavior.

[1]: https://docs.meteor.com/api/collections.html#Mongo-Collection-update