Add route for "Experiments"
This section will be for mini experiment projects, such as projects that
demonstrate use of a single class, or that work as simple tools for
specific tasks.

The aim of what's included isn't to be completely original, but to be a
collection of different ideas and tools built for the sake of
experimenting and learning -- and possibly ending up with something
that's useful later on (even if it's been done before).