Correct a parent selector error
Since these styles appeared inside of the parent selector of `+ span`,
there was no need to preceed them with a `+`.

This also revealed something about the behavior of `&`: it doesn't only
get replaced with the selector of the immediate parent's selector, but
a selector of the entire ancestry.

This was causing `+ &, + &:before` to compile to the selector below.

```
+ .project-manager-row input[type="checkbox"] + label,
+ .project-manager-row input[type="checkbox"] + label:before
```

Naturally, this selector is invalid since there's nothing before the
first `+` to indicate a sibling that `.project-manager-row` would have
been following after.

Lesson learned...