First attempt at setting up a test SMTP server locally
This uses MailDev as the test SMTP server. Throwing an error so far,
but hopefully this will be easy to resolve.

For some reason, it's behaving like STARTTLS should be used. Could be
the email package itself trying to force that... Will see if there's a
way to override that.